Complete first draft. Meets requirements
Program now meets the minimum requirements issued in the technical
challenge. Take the list of links leading to an image, read the list of
images and find 3 most prevalent colors in the RGB scheme in hexadecimal
formt (#000000 - # FFFFFF) in each image, and write the result into a
CSV file in the form url,color,color,color.