Add max-heap to extract colors
For main_test.go altered tests to support using a heap data structure
rather than having a map turned into a slice and then sorted. Changing
the map directly into a heap is O(n), the length of the map, rather than
O(n log n) for the sorting of the slice.

Also, removed TestGetImageData test since it could be tested in the
TestExtractURLs function by adding the error causing URLs to the test
input file, input_test.txt, rather than having a seperate test function.

Removed the conversion of the status check since if there's an error
the response status is provided in the error message.