Reduce number of YCbCr to RGB conversions
This update reduces the numbers of YCbCr color model conversions as
only the top 3 colors need to be in the RGB scheme. This takes the
number of conversion operations down from an O(m x n) amount for each
image where m and n are the horizontal and vertical number of pixels
to a constant amount with only 3 conversion calls for each image
processed.

This update has an issue with hanging. The waitgroup is waiting, but the
functions should be complete.